requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
ipykernel
tensorflow
keras
streamlit
opencv-python
pillow


